Add gradle, restructure project for gradle
