+++ .gitignore
@@ -22,12 +22,16 @@
 # IntelliJ
 .idea/workspace.xml
 .idea/tasks.xml
-out/
+/out/
 
 .gradle
 /build/
-bin/
+/bin/
 
 # Emacs
 *~
 \#*\#
+.\#*
+
+# OS X
+.DS_Store
