--- .idea/sqldialects.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/resources/sql.d/DatabaseSchema.sql" dialect="Derby" />
-  </component>
-</project>
