+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Wed Mar 29 20:51:11 EDT 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-3.4.1-bin.zip
