+++ src/main/java/userpanel/Controller.java
@@ -14,14 +14,14 @@ public class Controller {
 
     @FXML
     private void logAsAdminClicked() throws Exception{
-//        Parent root = (AnchorPane)FXMLLoader.load(getClass().getResource("../resources/AdminUI.fxml"));
+//        Parent root = (AnchorPane)FXMLLoader.load(getClass().getResource("/AdminUI.fxml"));
 //        adminScene.setTitle("Faulkner Hospital Navigator");
 //        primaryStage.setScene(new Scene(root, 1174, 722));
 //        primaryStage.setScene(adminStage);
 //        adminScene.show();
 
         Parent root;
-        root = FXMLLoader.load(getClass().getResource("../resources/AdminUI.fxml"));
+        root = FXMLLoader.load(getClass().getResource("/AdminUI.fxml"));
         Stage stage = new Stage();
         stage.setTitle("My New Stage Title");
         stage.setScene(new Scene(root, 1174, 722));
